move class ToastMessage out from ChromeViewModel
