+++ app/src/main/java/org/mozilla/rocket/util/ToastMessage.kt
@@ -1,0 +1,12 @@
+package org.mozilla.rocket.util
+
+class ToastMessage(
+    val stringResId: Int,
+    val duration: Int = LENGTH_SHORT,
+    vararg val args: String
+) {
+    companion object {
+        const val LENGTH_SHORT = 0
+        const val LENGTH_LONG = 1
+    }
+}
